+++ controller/owner.js
@@ -6,6 +6,8 @@ export const getOwner = async (req,res,next) =>{
         const id = req.params.id
         const user = await ownerService.getOwner(id)
         if(!user) return res.status(404).json({msg: "User Not Found", status: false})
+        if(id != req.user.userId) 
+            return res.status(404).json({msg: "Fail", status: false})
         res.status(200).json({msg:"User Found", data: user, status: true})
     } catch (error) {
         console.log(error)
@@ -16,7 +18,9 @@ export const getOwner = async (req,res,next) =>{
 export const updateOwner = async(req,res,next) => {
     try {
         const {phone,password, name, string, birthday, address, email} = req.body
-        const id = req.user.userId
+        const id = req.params.id
+        if(id != req.user.userId) 
+            return res.status(404).json({msg: "Fail", status: false})
         const hashPassword = await bcrypt.hash(password,8)
         const updatedUser = await ownerService.updateProfile({phone, password:hashPassword, name, string, birthday, address, email,id})
         if(!updatedUser) return res.status(404).json({msg: "User Not Found", status: false})
@@ -29,7 +33,10 @@ export const updateOwner = async(req,res,next) => {
 
 export const deleteOwner = async(req,res,next) => {
     try {
-        const id = req.user.userId
+        const id = req.params.id
+        console.log(req.user.userId, id)
+        if(id != req.user.userId) 
+            return res.status(404).json({msg: "Fail", status: false})
         const deletedUser = await ownerService.deleteOwner(id)
         if(!deletedUser) return res.status(404).json({msg: "User Not Found", status: false})
         res.status(200).json({msg:"Delete success", data: deletedUser, status: true})
