+++ controller/room.js
@@ -1,8 +1,13 @@
+import houseService from '../services/houseService'
 import roomService from '../services/roomService'
 
 export const createRoom = async (req,res,next) => {
     try {
         const {name, house_id,max_user, description,status, cost } = req.body
+        const house = await houseService.getHouse(house_id)
+        const isBelong = house.owner_id == req.user.userId
+        if (!isBelong) 
+            return res.status(404).json({msg: "Fail", status: false})
         const createdRoom = await roomService.createRoom({name, house_id,max_user, description,status, cost})
         res.status(200).json({msg:"Created success", data: createdRoom, status: true})
     } catch (error) {
@@ -26,6 +31,10 @@ export const getRoom = async (req,res,next) =>{
     try {
         const id = req.params.id
         const room = await roomService.getRoom(id)
+        const house = await houseService.getHouse(room.house_id)
+        const isBelong = house.owner_id == req.user.userId
+        if (!isBelong) 
+            return res.status(404).json({msg: "Fail", status: false})
         if(!room) return res.status(404).json({msg: "Not Found", status: false})
         res.status(200).json({msg:"Found", data: room, status: true})
     } catch (error) {
@@ -38,6 +47,11 @@ export const updateRoom = async(req,res,next) => {
     try {
         const  {name, house_id,max_user, description,status, cost } = req.body
         const id = req.params.id
+        const room = await roomService.getRoom(id)
+        const house = await houseService.getHouse(room.house_id)
+        const isBelong = house.owner_id == req.user.userId
+        if (!isBelong) 
+            return res.status(404).json({msg: "Fail", status: false})
         const updatedRoom = await roomService.updateRoom({name, house_id, max_user, description,status, cost, id})
         if(!updatedRoom) return res.status(404).json({msg: "Not Found", status: false})
         res.status(200).json({msg:"Updated", data: updatedRoom, status: true})
@@ -50,6 +64,11 @@ export const updateRoom = async(req,res,next) => {
 export const deleteRoom = async(req,res,next) => {
     try {
         const id = req.params.id
+        const room = await roomService.getRoom(id)
+        const house = await houseService.getHouse(room.house_id)
+        const isBelong = house.owner_id == req.user.userId
+        if (!isBelong) 
+            return res.status(404).json({msg: "Fail", status: false})
         await roomService.deleteRoom(id)
         res.status(200).json({msg:"Deleted", status: true})
     } catch (error) {
